Make debug source correct, even on low debug or for block compile.
This essentially goes back to the state of affairs before the change
5c2ad7a3fa3d3ee2d55b0b04823ecd8f21944898. That change was completely
reasonable at the time because there was no way to have a compiled fun
be associated with more than one TLF number. Now that we have block
compilation and auto tlf merging working again, we need to store the
TLF number at a finer granularity again, and to also record the
START-POSITIONS vector, which at least obviates the need for storing
the char offset in each compiled debug object. One interesting thing
that the restored implementation does is that if each code location in
a function shares the same tlf number, that is taken into account and
we get no overhead in the fasl in that case.

One difference is that instead of incrementally building up the
START-POSITIONS vector on each component load, which takes up a lot of
space and is annoying to do, now we just forget about using the
character offset if loading gets interrupted, and just back up and
read the forms out of the source info. This makes debug information
during interrupted loads work, without all the extra bookkeeping.

The end result is that debug source works better for debug 1 and for
block compilation than previous releases, and we also fix it up for
high debug after it got worse from adding top-level lambda merging.