Simplify zero_dirty_pages
There are two approaches, both easier than before:
* for #+darwin-jit, if any page in [start,end] has need_to_zero
  then call zero_range_with_mmap() once for the whole range.
* for #-darwin-jit, call memset on a per-page basis

In neither case do we operate on a strict subrange of [start,end]
that is less than the entire range but more than one page.
Removing the inner loop makes perfect sense because >90% of all
requests to clear dirty pages are for a single page, and in the case
of (MAKE-ARRAY 1000000 :INITIAL-ELEMENT NIL), the allocator needs
to be redesigned to perform NIL-fill instead of 0-fill anyway.