Fix function type clobbering in a principled way.
The test case that brought on FUNCTIONAL-TOP-LEVEL-DEFUN-P was always
a bit suspect because we've decided that by default we want to assume
functions in the same file can't be redefined, and hence the test case
should have undefined behavior. The fix introducing
FUNCTIONAL-TOP-LEVEL-DEFUN-P (9cb20dfbaa0dc4e1f51c8b4d5ed5601f4482619b)
was also just wrong because it broke another test case that was
basically the inverse of the original test case. So we
extend *DERIVE-FUNCTION-TYPES* to actually be able to handle the case
when people want the full ability to do function redefinition, and
leave *DERIVE-FUNCTION-TYPES* to :SAME-FILE (a new keyword) by
default, which matches the status quo behavior.

This area of the compiler could probably use some clean up, since it
seems in many places we were two minds when it came to how much of the
early binding in the same file we actually wanted to expose type-wise
in the compiler.