Remove size_immediate
That function was really "size_cons" which should have returned a constant
value of 2 instead of 1. I think in years past it was required to return 1
because reasons. Rather than change its behavior, we just don't need it
since OBJECT_SIZE prechecks for conses anyway. But we can actually do that
test more efficiently by storing a sentinel value of 0 in sizetab[].
If the sizer is 0, then the object is a cons, otherwise call the sizer.

I intend to rename OBJECT_SIZE as it is no longer a preprocessor macro,
but that's a search-and-replace which touches a lot of lines,
and I'd like this change to pass muster independently of that.