Improve atomic logging
Rule #1 for logging during signal handling is never call anything
in the printf() family of functions, as they might malloc or
otherwise be somehow nonreentrant. i.e. FSHOW is just wrong.
(And the second rule is never do something to induce so much delay
as to make the problem you're investigating irreproducible)