Fix variable type proclamation scoping with block compilation.
To do this, we need to start IR1 converting all special variable
references into REFs again. We changed this in the past to avoid
eliding the boundp check, but this can be simply done by making sure
not to flush references to potentially unbound variables while
flushing dead code, since that's the only place we can in fact flush
dead code.

That obviates the need to have separate SYMEVAL and SYM-GLOBAL-VAL
aliases, so remove them as well.