x86-64: Untangle print-mem-ref a bit
I hoped to gain some consistency in how references to the symbol's value
slot is printed, but boy is this code weird. Sometimes it prints
(SYMBOL-VALUE '*G*) and sometimes it prints just *G*. The newest bug
was that we got two end-of-line notes about one symbol with an intervening
newline, because of code guarded by #+sb-thread that shouldn't have been,
so when I added "another" note, two notes came out for #+sb-thread.