Don't do FDEFN elision for DEFKNOWN of non-internal symbols.
Users shouldn't know or rely on this behavior, (they shouldn't be
using DEFKNOWN at all, really), but if they did, without this change,
redefinition would break without a NOTINLINE, and more perspicuously,
deferred tll evaluation during fasloading doesn't work, as DEFKNOWN
tells the compiler that the function definition will in fact always
have been defined. So disable this for non-internal functions.

I think in general, FDEFN elision is probably only warranted in very
few situations, if at all. DEFKNOWN really tells the compiler: This is
what the compiler is allowed to assume about calls to functions, not
that the definition of the function can never change or that the
definition of the function will even always be defined.