Fix DELETE-PACKAGE interaction with optimistic symbol loading.
Instead of NILing out the symbol packages of symbols inside deleted
packages while fasloading, install a cookie that gets upgraded into a
real package if the package gets redefined again. Since DELETE-PACKAGE
causes SYMBOL-PACKAGE of the package's symbols to be an
implementation-defined value, it's OK to install something non-NIL. If
the cookie doesn't get upgraded to a real package by the end of the
load, it will get GC'd when the rehoming package graph goes out of
scope.