defconstant: ignore equivalent SAPs.
SAPs behave like numbers or characters and can be copied at any time,
yet EQL doesn't work on them. Making EQL work on SAPs is fairly
complicated and there are many implicit assumptions about what can be
compared with EQ or EQL. And it's unclear how significant a deviation
from the standard that would be.

And nobody will notice that the constant will not be EQ to its
previous value, as it already can be copied at any time.

Reported by Bart Botta.