Remove a lot of #+- from generic/type-vops
I didn't like the amount of architecture-specific code in the generic file
such as SIMPLE-TYPE-PREDICATE which is specific to x86-64.
Now we define the architecture-specific type-vops first, then the
fallback definitions, and avoid overwriting any definition with
a fallback if there was alread something there. It's not required
to explicitly avoid making the generic definition, it'll do the right
thing. (In fact it may never have been "required" to have the conditionals
as evidenced by the typo at "#-arm-64". We'd just stomp on the generic
definition with the one from the backend, which is onfusing in itself)