Use zstd for core compression instead of zlib.
It is much faster, and compression level 1 with zstd (whose max level
goes up to 22) produces smaller cores than compression level 9 with
zlib.

The default value of 9 for zstd compression level was chosen because
it compresses a fresh warm core better than the old zlib default of 1
while doing it faster.