Disable tests of breakpoint-based tracing on ppc and ppc64
We hadn't one regression test of multiple-value returns with
breakpoints, so I added one, which failed on ppc.
We know already that breakpoints don't work on freebsd,
so just stop testing them entirely.

Even though single-value returns seemed to work, the debugger
is healthier by not having to cater to already-on-life-support
logic. Specifically, we can remove one variation on mapping a
fun-end breakpoint to its metadata: we don't need the
*FUN-END-COOKIES* table because the so-called bpt-LRA
(formerly "bogus LRA") can have as many slots as you like;
but as git rev e7dabe2d said, I couldn't figure it out for ppc.