Give more flexibility to the code allocator
This change will lend itself to an improvement in the immobile code allocator
by allowing it to oversize a new object in case it is unable to split a hole
whose size is below a threshold. Assume that it can split a hole with 4 words
left over (not that those can be used), but will NOT split leaving just 2 words.
In that case, the code object absorbs 2 words. The overage has to be accounted
for in the new object since there is no other place to store a physical size.