Contribs can now be disabled individually with --without-CONTRIB.
This mechanism allows developers to selectively disable the build of
certain contribs.  Each --without-CONTRIB flag supplied to make.sh ends up
in the CONTRIB_BLOCKLIST variable that is part of the build configuration.
Each contrib that is mentioned in that blocklist is neither built nor
tested.

I hope this mechanism makes the enormous compile times of my sb-simd
contrib a bit more tolerable :)