take contrib blocklist into account for doc generation
If some contrib module is blocked for building also block it in
documentation generation since this will fail otherwise.

Additionally this fixes a shift operation that is needed
unconditionally when parsing the script parameters since otherwise
the first parameter is used for the second parameter as well if it
turns out to be empty.

committer: Christophe Rhodes <csr21@cantab.net>