Simplify & fix a bug in immobile symbol allocator
The C fallback didn't adjust the free_index if it had just consumed the highest
addressable object on a page. That helped my first attempt at the allocation vop
which wanted the free_index to always be a usable object. But it didn't work, so
the vop checks for overrun prior to using the index in any case. That being so,
free_index can and must be bumped up, otherwise update_immobile_nursery_bits()
might not see a newly allocated object. To fix it, get rid of two slots from the
page struct, and always reflect the correct generation mask right away.