Restore undefined variable warnings
The change in d788babebfef30af6ff0a329865de20b849f749f to respect
scope of mufflings inadvertently removed warnings from a number of
cases, including type and extent declarations and setq forms.  Restore
them, fixing in the sources a few instances of undefined declarations
that have crept in, and an instance in our tests whose behaviour
inexplicably changed somehow.  This also changes the condition kind of
a free, unknown dynamic extent declaration on a variable to a full
warning (it was a style warning before).