Revise crusty old gencgc_apply_code_fixups() function
The comments implied that we store a vector of fixups, but technically it has
been a packed integer ever since rev f7c8fe8142. And years ago the logic
was flaky (possibly harming the object being fixed), but that bug was resolved
by rev f76d666a2c so remarks about corruption are misplaced.
var-io can assert that its input is valid as a reasonable precaution.