x86-64: give all funcallable-instances a builtin trampoline
For a mere 2 words, they're all callable using the same convention as
a simple-fun, i.e. the entry address can be assigned into the raw-addr
slot of an fdefn with no fuss.  It does not matter whether they're in
immobile space - in fact they aren't. But the feature is predicated on
having immobile fdefns because that's what allows a 1-instruction call.