Fix regression from 9dae557697
That change failed to recognize a critical distinction between FDEFNs used
for global #'FUN reference versus named-call. If an FDEFN was used for both
in one caller, crashes might result if static linking caused the FDEFN in
a code header to be replaced by a CONS listing the PC locations, because
the #'FUN use would see a CONS instead of an FDEFN (due to rev 99c2a8e853).
So stop replacing ambiguous cases. I prefer this over having to create
metadata to remembers why and where a particular FDEFN was replaced.