Revise blt-copier-for-widetag yet again
Change 88d18542ad wanted to avoid having to know when global constants are
dumped, but if we're going to define a function to solve such problem,
it's not much better to require that you call a new cold-init thunk prior to
calling the new function. And it retained 1kb of junk because of the closure.