Fix incredibly rare GETHASH bug on unsynchronized table
The worst-case-fallback algorithm for concurrent GETHASH was basically right,
but returned a "random" element due to an off-by-2x error on the k/v index.

This bug would only be possible to observe under the following scenario:
- the table hashed some keys by address and some by their contents
- the need-to-rehash flag was set by GC but not observed in Lisp yet
  (as is always the case after save-lisp-and-die in a freshly started core)
- GETHASH is started in two threads nearly simultaneously
- the thread losing the race to rehash seeks a key *NOT* hashed by address

In any of the preceding conditions are unmet, then the code path taken
would utilize correctly-working logic, not the buggy logic.