Delete irrelevant FIXME from initial commit of SBCL
A "FIXME" here is weird, because it's not saying what to fix; it's simply saying
what exists. I think the implication was that somebody didn't like the declared
slot type in CMUCL of single-float because technically it doesn't preserve the
user's request. We're free to ignore the argument entirely though.