Fix bug235a completely differently.
Commit 7c406887c08477181e869b1b98142d99b52990ac fixed bug235a by
turning DEFINED-FUN-FUNCTIONAL into a list of functionals for each
possible policy. Nikodemus himself says in launchpad that this is the
wrong solution, because it doesn't really make sense to do it this way
with block compilation. Indeed, with block compilation we had to
kludge finding the correct defined-fun-functional and just grab the
first one when that was definitely not the right thing to do.

The solution in the aforementioned comment was fundamentally wrong
because it misdiagnosed the problem: it's not that inline expansions
only shared one functional and hence we needed more of them per policy
- they in fact were never getting reused as is -- it was just used to
hold the inline expansion and lexenv, which was what was being used to
convert a fresh functional per call-site via local analysis (done this
way to I think make local function inlining share the same machinery
as global function inlining), but the problem actually was just that
MAYBE-EXPAND-LOCAL-INLINE wasn't using the call site policy when
copying and local-inlining the inline expansion stored in the
functional, instead using the policy in the saved lexenv.

So instead of converting and storing different functionals for each
policy in the DEFINED-FUN (reconverting again at each call site taking
the inline expansion and lexenv from the assortment of
functionals-whose-only-difference-is-lexenv-policy), we just store one
functional and reconvert again for any new copies with the policy of
the call site. We end up converting the exact same number of inline
expansions without the extra space overhead of a list of functionals,
so the end result is we have a simpler and more memory efficient
solution to the same problem that also just works with block
compilation picture, which is good because DEFINED-FUN was designed in
the first place for code compiled in the same block compilation.