Hash-cons alien-type representations at compile-time
This has no effect on fasls unfortunately because the load-form is
the default. So a typical application may end up with 100 instances
of alien-type (signed-byte 8) in the core. To fix that, a custom
load form is needed. But the caches "mostly" work, except for where
they (or I) get confused because of mutable instances.

Also: use fewer DEF!STRUCTs.