Hash-cons all ctypes
This gets rid of the coalescing step in save-lisp-and-die as well as
several ad-hoc caches such as the cache formerly known as
unknown-type-atoms, and the values-type-cached function.
Also remove a few special cases of "interned" types since basically
everything looks interned now.

Slower constructors are more than made up for by improved hit rates
in type operator caches. In fact, the compiler ran about 3% faster
in several benchmarks, including a timed run of make-host-1.

And at last it should be reasonably ok to execute on an arena without
fear that every TYPEP operation may create a heap->arena reference due
to caches holding on to type specifiers and internal representations.