Simplify FIND-CONSTANT (again).
* Remove the NAMED-CONSTANTS table. It is entirely redudant since the
FREE-VARS table exists.

* Remove some outdated comments that predate the use of a "real"
similarity table for the IR1 constant coalescing stuff, and add a
comment explaining the real point of constant coalescing in IR1 (as
opposed to in the dumper).

* Call FIND-CONSTANT only when we want an anonymous value. Named
constants that should be dumped as such don't go through FIND-CONSTANT
anymore. They get the same constant LEAF through the FREE-VARS table,
as mentioned already.

* As a result, *IR1-NAMESPACE* is now always bound. Don't
conditionalize on its boundedness anymore.

* Make named constants with any number value dump anonymously, even
when file-compiling. Modify a test to reflect that reality by using
EQL instead of EQ, and cite the CLHS glossary to prove that named
constant references *do* need to always yield EQL values (to other
references of the same named constant). Hopefully that clears up a
massive source of confusion, because many seem to interpret the spec
as not mandating that.