Define COMPILED-FUNCTION more in accordance with its intent
To avoid claiming that interpreted methods of generic-functions are
compiled, conservatively deem that no generic-function is compiled.

A COMPILED-FUNCTION has clear constraints about what should be true of
calls to it: it will not invoke macroexpanders nor cause LOAD-TIME-VALUEs
to be computed. It's perfectly ok for functions that do *not* satisfy
COMPILED-FUNCTION-P to meet those requirements, not the other way around.

Incidentally, fix lp#1537003