Implement use-after-rewind detection in arenas
not exactly use-after-free because it does not free the memory,
but it deals with a similar issue. By changing REWIND-ARENA to HIDE-ARENA,
and then working with a different arena, you can trap erroneous references
to the hidden arena without the further damage that would come from
seeing partially overwritten memory if you reused that same arena.