Remove sb-graph.
It's bitrotted already, and from the dot files produced in the example
directory it seems like there's quite a long way to go before it
produces anything readable. The function IR1-TO-DOT produces dot files
based directly on the normal trace file output that people use to
understand IR1, so it seems better just to remove sb-graph, which also
takes extra time to compile.