Remove the policy CHECK-TAG-EXISTENCE.
It's not exported, and we don't need it because any uses of it should
imply that the local function containing the exits should be declared
DYNAMIC-EXTENT. (Code in which the exits are only reached in dynamic
extent uses of local functions are, well, naturally unmaintainable and
unsafety prone). So remove remaining uses of the policy and delete
it. Keep the facet that SAFETY 0 turns off all extent checking though,
as noted in the failure note.