Adjust %BITS slot usage in CTYPE instances
This makes the value no larger than a target fixnum, so it should be possible
to DX-allocate a temporary key in all type constructors regardless of whether
raw slots can go on the stack. Also %BITS can be directly used as the hash
for 1-arg cached functions such as TYPE-NEGATION, PRIMITIVE-TYPE-AUX, etc.

Additionally, reserve the low 2 bits to use as CONTAINS-UNKNOWN and
CONTAINS-HAIRY flags, as I plan to compute them exactly once per instance.