Compute and store CONTAINS-UNKNOWN-TYPE-P for all ctypes
(and CONTAINS-HAIRY-TYPE-P) by taking 2 bits of the %BITS slot.

In doing so I discovered a discrepancy with the former implementations
based on MAP-TYPE, which does not descend into CONSTANT-TYPE.
Initially I thought it must be inadvertent, because the
"(:constant (satisfies power-of-two-limit-p))" terms in type-vops
produced a mismatch in my new-versus-old cross-check, but when I tried
changing MAP-TYPE to examine CONSTANT-TYPE-TYPE, it caused test failures.
That seemed super weird. Perhaps some tests are wrong.

At any rate, we're stuck for the time being with a new way and an old way
to define CONTAINS-UNKNOWN, and the new works as intended in OK-TO-MEMOIZE-P.