Handle top level DX closures correctly.
We need to accumulate into the right component's dx-lvars list. The
component of the ENCLOSE node might be different from the component we
are currently scanning in, so fix that. Otherwise we get a bad
reference to another component when grovelling the list during stack
analysis. (We could probably avoid this by making the DX plumbing
sane, but that's not easy to do given our DX representation.)