Add ability to intercept SETF of LEAF-TYPE etc
This is useful for finding things like bad memoization and/or type derivation.
You can stop the compiler at the just the right place by injecting handcrafted
expressions that detect a leaf with a certain name and type.
(I needed this to track down the problem that rev 2c2bd92839 fixed)

There's no runtime impact if you leave the INLINE declamations intact.
But if we don't like this it could aways be reverted after figuring out
why the DCE tests seem to have platform-specific failures.