Speed up FINALIZE and CANCEL-FINALIZATION w/ many threads
This replaces the usual hash-table with a new storage representation
that is almost always lockfree, and removes all complexity around
assigning a unique ID to each finalized object.
The ugly hack in cull_weak_hash_table_bucket() is no more.

As the benchmark shows, with 4 threads we can create finalizers
about 3x faster at the cost of about 1.5x more memory.

Fixes lp#1998064 where finalizers are concerned, but if the bug
exists more generally, this doesn't do anything for it.