assignment-convert: Check destination lvar is also the same.
Normally this is true by virtue of the ctrans being the same, as there
is a tight bundling between lvars and ctrans in normal code by virtue
of them being facets of the same continuation. However this invariant
can actually be violated by code deletion caused by optimizations, so
to stay on the safe sode we check for the value part of the
continuations being the same as well across outside calls

Fixes lp#2002310.