Avoid using FIND-NAMED-CALLEES in tests
Use the new CTU:IR1-NAMED-CALLS function instead because:
* It's easy to assert that the called names EQUAL a list of symbols.
* I plan to skip emitting FDEFNs for certain symbols (SB-KERNEL: and CL: etc)
  on x86-64 if not other architectures. Whether an fdefn exists would be
  unknown to the test (it only works for pseudostatic fdefns, e.g.)

Alternatively, ASM-SEARCH may be helpful in searching for a string.
In any event, scaning the code header will not be reliable.

Additionally, try to express test expectations as positive assertions
which is robust against accidentally failing to find something for a wrong
reason like a flawed test. (inline signum was one such)