Make redblack a pure test
It basically was because we allow package creation in pure tests.
Just do a better job of scrubbing changes to PRINT-OBJECT.
(The purely functional redblack algorithm isn't built in any more,
but it's nice to have as another GC stress test)

Not to mention, the test manifest file was incorrect, because once
it was a pure test, then it wasn't. And apparently we have a bug that
the scaffolding doesn't check impure requisites carefully enough.