Delete NOTE-LOCAL-FUNCTIONAL, removing a KLUDGE.
Commit 36a379d746b9eb74ba8c5afff40dc5dcb9f4557a introduced this
function to work around issues of referencing already converted local
functionals. Turns out this was only needed because the IR1 %DEFUN
translator was rewritten incorrectly when it became the
EVAL-WHEN (:COMPILE-TOPLEVEL ...) style of top level processing. (And
note that CMU CL didn't have this problem). The named lambda machinery
meant to replace the IR1 translator is a weird twisted duplicated
version of IR1-CONVERT-INLINE-EXPANSION, and handling the
DEFINED-FUNS/FREE-FUNS stuff became wrong. Fix the discrepancy,
allowing us to get rid of NOTE-LOCAL-FUNCTIONAL. We still essentially
have two duplicated code paths though. It would be nice to reunify
them again like they were originally before the IR1 translator for
%DEFUN was removed.

ALso, make sure to nuke the ir1-namespace after we finish compiling a
component. That is now another way to have dangling references to
already converted functionals.

Move ftype clearing from %COMPILER-DEFUN to named lambda processing:
THis way we still warn about self call argument type conflicting
without messing with GET-DEFINED-FUN. This also tackles 2 FIXMEs: It's
clear that line has nothing to do with block compilation and
everything to do with ftypes for redefined functions. It's also now
clear that asserting the proclaimed type is orthogonal to whether we
have a non-entry block compiled function.

The upshot of this change is that now it's easier to see that we are
actually treating local functionals correctly with block compilation
on. Before, it just seemed like we were doing an ad-hoc workaround
that might expose the same bug in NOTE-LOCAL-FUNCTIONAL with it turned
on.

Incidentally fix a test that abused named-lambda: It was only working
coincidentally because we got the logic for functional saving wrong:
It was recognizing the self call even without the recognize-self-calls
declaration.