Remove some unreachable code in IR2-CONVERT-CLOSURE.
Commit 46602bb31b943b1793da732781586c032333c907 added a workaround for
emitting a reference to a closure that's actually an inlinable global
function, and redericting it to refer to the global definition
instead. We now avoid this problem in IR1 by pointing references to a
converted functional only when it was actually defiend in the same
compilation block, so we don't need to check for this case in IR2
anymore, as it was just a workaround which didn't help with the object
identity problem.