Stop blowing away all symlinks in clean.sh
make.sh should not have the default behavor of destroying work-in-progress
involving symlinks. The "solution" of creating a file of precious names was
not very discoverable and I only stumbled upon it after I thought I saw
symlinks disappear several times in a row, before I realized that I was not
imagining it.